groot/riofs: consolidate compression into groot/internal/rcompress
