+++ groot/riofs/compress.go
@@ -5,55 +5,21 @@
 package riofs
 
 import (
-	"bytes"
 	"compress/flate"
-	"compress/zlib"
-	"encoding/binary"
-	"io"
 
-	"github.com/pierrec/lz4"
-	"github.com/pierrec/xxHash/xxHash64"
 	"github.com/pkg/errors"
-	"github.com/ulikunitz/xz"
+	"go-hep.org/x/hep/groot/internal/rcompress"
 )
 
-type compressAlgType int
-
-// constants for compression/decompression
-const (
-	kZLIB                          compressAlgType = 1
-	kLZMA                          compressAlgType = 2
-	kOldCompressionAlgo            compressAlgType = 3
-	kLZ4                           compressAlgType = 4
-	kUndefinedCompressionAlgorithm compressAlgType = 5
-)
-
-var (
-	// errNoCompression is returned when the compression algorithm
-	// couldn't compress the input or when the compressed output is bigger
-	// than the input
-	errNoCompression = errors.New("riofs: no compression")
-)
-
-// Note: this contains ZL[src][dst] where src and dst are 3 bytes each.
-const rootHDRSIZE = 9
-
-// because each zipped block contains:
-// - the size of the input data
-// - the size of the compressed data
-// where each size is saved on 3 bytes, the maximal size
-// of each block can not be bigger than 16Mb.
-const kMaxCompressedBlockSize = 0xffffff
-
-func (f *File) setCompression(alg compressAlgType, lvl int) {
+func (f *File) setCompression(alg rcompress.Kind, lvl int) {
 	switch {
 	case lvl == flate.DefaultCompression:
 		switch alg {
-		case kLZ4:
+		case rcompress.LZ4:
 			lvl = 1
-		case kLZMA:
+		case rcompress.LZMA:
 			lvl = 1
-		case kZLIB:
+		case rcompress.ZLIB:
 			lvl = 6
 		default:
 			panic(errors.Errorf("riofs: unknown compression algorithm: %v", alg))
@@ -67,7 +33,7 @@ func (f *File) setCompression(alg compressAlgType, lvl int) {
 // WithLZ4 configures a ROOT file to use LZ4 as a compression mechanism.
 func WithLZ4(level int) FileOption {
 	return func(f *File) error {
-		f.setCompression(kLZ4, level)
+		f.setCompression(rcompress.LZ4, level)
 		return nil
 	}
 }
@@ -75,7 +41,7 @@ func WithLZ4(level int) FileOption {
 // WithLZMA configures a ROOT file to use LZMA as a compression mechanism.
 func WithLZMA(level int) FileOption {
 	return func(f *File) error {
-		f.setCompression(kLZMA, level)
+		f.setCompression(rcompress.LZMA, level)
 		return nil
 	}
 }
@@ -91,265 +57,7 @@ func WithoutCompression() FileOption {
 // WithZlib configures a ROOT file to use zlib as a compression mechanism.
 func WithZlib(level int) FileOption {
 	return func(f *File) error {
-		f.setCompression(kZLIB, level)
+		f.setCompression(rcompress.ZLIB, level)
 		return nil
 	}
-}
-
-func rootCompressAlg(buf [rootHDRSIZE]byte) compressAlgType {
-	switch {
-	case buf[0] == 'Z' && buf[1] == 'L':
-		return kZLIB
-	case buf[0] == 'X' && buf[1] == 'Z':
-		return kLZMA
-	case buf[0] == 'L' && buf[1] == '4':
-		return kLZ4
-	case buf[0] == 'C' && buf[1] == 'S':
-		return kOldCompressionAlgo
-	default:
-		return kUndefinedCompressionAlgorithm
-	}
-}
-
-func rootCompressAlgLvl(v int32) (compressAlgType, int) {
-	var (
-		alg = compressAlgType(v / 100)
-		lvl = int(v % 100)
-	)
-
-	return alg, lvl
-}
-
-func compress(compr int32, src []byte) ([]byte, error) {
-	const (
-		blksz = kMaxCompressedBlockSize // 16Mb
-	)
-
-	alg, lvl := rootCompressAlgLvl(compr)
-
-	if alg == 0 || lvl == 0 || len(src) < 512 {
-		// no compression
-		return src, nil
-	}
-
-	var (
-		nblocks = len(src)/blksz + 1
-		dst     = make([]byte, len(src)+nblocks*rootHDRSIZE)
-		cur     = 0
-		beg     = 0
-		end     = 0
-	)
-
-	for beg = 0; beg < len(src); beg += blksz {
-		end = beg + blksz
-		if end > len(src) {
-			end = len(src)
-		}
-		// FIXME(sbinet): split out into compressBlock{Zlib,LZ4,...}
-		n, err := compressBlock(alg, lvl, dst[cur:], src[beg:end])
-		switch err {
-		case nil:
-			cur += n
-		case errNoCompression:
-			return src, nil
-		default:
-			return nil, err
-		}
-	}
-
-	return dst[:cur], nil
-}
-
-func compressBlock(alg compressAlgType, lvl int, tgt, src []byte) (int, error) {
-	// FIXME(sbinet): rework tgt/dst to reduce buffer allocation.
-
-	var (
-		err error
-		hdr [rootHDRSIZE]byte
-		dst []byte
-
-		srcsz = int32(len(src))
-		dstsz = srcsz
-	)
-
-	switch alg {
-	case kZLIB:
-		hdr[0] = 'Z'
-		hdr[1] = 'L'
-		hdr[2] = 8 // zlib deflated
-		buf := new(bytes.Buffer)
-		buf.Grow(len(src))
-		w, err := zlib.NewWriterLevel(buf, lvl)
-		if err != nil {
-			return 0, errors.Wrapf(err, "riofs: could not create ZLIB compressor")
-		}
-		_, err = w.Write(src)
-		if err != nil {
-			return 0, errors.Wrapf(err, "riofs: could not write ZLIB compressed bytes")
-		}
-		err = w.Close()
-		if err != nil {
-			return 0, errors.Wrapf(err, "riofs: could not close ZLIB compressor")
-		}
-		dstsz = int32(buf.Len())
-		if dstsz > srcsz {
-			return 0, errNoCompression
-		}
-		dst = append(hdr[:], buf.Bytes()...)
-
-	case kLZMA:
-		hdr[0] = 'X'
-		hdr[1] = 'Z'
-		cfg := xz.WriterConfig{
-			CheckSum: xz.CRC32,
-		}
-		if err := cfg.Verify(); err != nil {
-			return 0, errors.Wrapf(err, "riofs: could not create LZMA compressor config")
-		}
-		buf := new(bytes.Buffer)
-		buf.Grow(len(src))
-		w, err := cfg.NewWriter(buf)
-		if err != nil {
-			return 0, errors.Wrapf(err, "riofs: could not create LZMA compressor")
-		}
-
-		_, err = w.Write(src)
-		if err != nil {
-			return 0, errors.Wrapf(err, "riofs: could not write LZMA compressed bytes")
-		}
-
-		err = w.Close()
-		if err != nil {
-			return 0, errors.Wrapf(err, "riofs: could not close LZMA compressor")
-		}
-
-		dstsz = int32(buf.Len())
-		if dstsz > srcsz {
-			return 0, errNoCompression
-		}
-		dst = append(hdr[:], buf.Bytes()...)
-
-	case kLZ4:
-		hdr[0] = 'L'
-		hdr[1] = '4'
-		hdr[2] = lz4.Version
-
-		const chksum = 8
-		var room = int(float64(srcsz) * 2e-4) // lz4 needs some extra scratch space
-		dst = make([]byte, rootHDRSIZE+chksum+len(src)+room)
-		buf := dst[rootHDRSIZE:]
-		var n = 0
-		switch {
-		case lvl >= 4:
-			if lvl > 9 {
-				lvl = 9
-			}
-			n, err = lz4.CompressBlockHC(src, buf[chksum:], lvl)
-		default:
-			ht := make([]int, 1<<16)
-			n, err = lz4.CompressBlock(src, buf[chksum:], ht)
-		}
-		if err != nil {
-			return 0, errors.Wrapf(err, "riofs: could not compress with LZ4")
-		}
-
-		if n == 0 {
-			// not compressible.
-			return 0, errNoCompression
-		}
-
-		buf = buf[:n+chksum]
-		dst = dst[:len(buf)+rootHDRSIZE]
-		binary.BigEndian.PutUint64(buf[:chksum], xxHash64.Checksum(buf[chksum:], 0))
-		dstsz = int32(n + chksum)
-
-	case kOldCompressionAlgo:
-		return 0, errors.Errorf("riofs: old compression algorithm unsupported")
-
-	default:
-		return 0, errors.Errorf("riofs: unknown algorithm %d", alg)
-	}
-
-	if dstsz > kMaxCompressedBlockSize {
-		return 0, errNoCompression
-	}
-
-	hdr[3] = byte(dstsz)
-	hdr[4] = byte(dstsz >> 8)
-	hdr[5] = byte(dstsz >> 16)
-
-	hdr[6] = byte(srcsz)
-	hdr[7] = byte(srcsz >> 8)
-	hdr[8] = byte(srcsz >> 16)
-
-	copy(dst, hdr[:])
-	n := copy(tgt, dst)
-
-	return n, nil
-}
-
-func decompress(r io.Reader, buf []byte) error {
-	var (
-		beg    = 0
-		end    = 0
-		buflen = len(buf)
-	)
-
-	for end < buflen {
-		var hdr [rootHDRSIZE]byte
-		_, err := io.ReadFull(r, hdr[:])
-		if err != nil {
-			return errors.Wrapf(err, "riofs: could not read compress header")
-		}
-
-		srcsz := (int64(hdr[3]) | int64(hdr[4])<<8 | int64(hdr[5])<<16)
-		tgtsz := int64(hdr[6]) | int64(hdr[7])<<8 | int64(hdr[8])<<16
-		end += int(tgtsz)
-		lr := &io.LimitedReader{R: r, N: srcsz}
-		switch rootCompressAlg(hdr) {
-		case kZLIB:
-			rc, err := zlib.NewReader(lr)
-			if err != nil {
-				return errors.Wrapf(err, "riofs: could not create ZLIB reader")
-			}
-			defer rc.Close()
-			_, err = io.ReadFull(rc, buf[beg:end])
-			if err != nil {
-				return errors.Wrapf(err, "riofs: could not decompress ZLIB buffer")
-			}
-
-		case kLZ4:
-			src := make([]byte, srcsz)
-			_, err = io.ReadFull(lr, src)
-			if err != nil {
-				return err
-			}
-			const chksum = 8
-			// FIXME: we skip the 32b checksum. use it!
-			_, err = lz4.UncompressBlock(src[chksum:], buf[beg:end])
-			if err != nil {
-				return err
-			}
-
-		case kLZMA:
-			rc, err := xz.NewReader(lr)
-			if err != nil {
-				return err
-			}
-			_, err = io.ReadFull(rc, buf[beg:end])
-			if err != nil {
-				return err
-			}
-			if lr.N > 0 {
-				// FIXME(sbinet): LZMA leaves some bytes on the floor...
-				lr.Read(make([]byte, lr.N))
-			}
-
-		default:
-			panic(errors.Errorf("riofs: unknown compression algorithm %q", hdr[:2]))
-		}
-		beg = end
-	}
-
-	return nil
 }
